finrisk/requirements.txt
# Core ML and Data Science
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
scipy==1.11.1

# Model Interpretability
shap==0.42.1
lime==0.2.0.1

# API and Web Framework
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3

# Database and Caching
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
redis==4.6.0
alembic==1.12.0

# ML Experiment Tracking
mlflow==2.6.0
optuna==3.3.0

# Monitoring and Metrics
prometheus-client==0.17.1
evidently==0.4.5

# Dashboard and Visualization
streamlit==1.26.0
plotly==5.15.0
dash==2.13.0
dash-bootstrap-components==1.4.2

# Data Processing
pyarrow==13.0.0
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
loguru==0.7.0
python-multipart==0.0.6
httpx==0.24.1
pytest==7.4.0
black==23.7.0
isort==5.12.0

# Financial and Risk
arch==6.2.0
statsmodels==0.14.0

# Security
cryptography==41.0.4
passlib==1.7.4

# Additional
joblib==1.3.2
tqdm==4.66.1
click==8.1.7 